Pulled out userInputForm into its own component.
